Bug Fixes to Profile Page, Review Display Changes, SearchClubs, ReviewDisplay and SignUpPage
I made some changes to routing of the profile page, users will need to update their display settings right after signing in with google or with email. Profile will only let you go to /Profile when you are
logged in. There are cosmetic label changes to the Clubs Form, and the searchbars in the review components are now fixed due to the additon of an import statement.